Only setup submodule/commit hook in Git checkout.

+++ autogen.sh
@@ -1,18 +1,21 @@
 #!/bin/sh
 
-# Make sure we have common
-if test ! -f common/gst-autogen.sh;
+if test -d .git;
 then
-  echo "+ Setting up common submodule"
-  git submodule init
-fi
-git submodule update
+  # Make sure we have common
+  if test ! -f common/gst-autogen.sh;
+  then
+    echo "+ Setting up common submodule"
+    git submodule init
+  fi
+  git submodule update
 
-# install pre-commit hook for doing clean commits
-if test ! \( -x .git/hooks/pre-commit -a -L .git/hooks/pre-commit \);
-then
-    rm -f .git/hooks/pre-commit
-    ln -s ../../common/hooks/pre-commit.hook .git/hooks/pre-commit
+  # install pre-commit hook for doing clean commits
+  if test ! \( -x .git/hooks/pre-commit -a -L .git/hooks/pre-commit \);
+  then
+      rm -f .git/hooks/pre-commit
+      ln -s ../../common/hooks/pre-commit.hook .git/hooks/pre-commit
+  fi
 fi
 
 if [ -x "`which autoreconf 2>/dev/null`" ] ; then
